requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
influxdb-client==1.38.0
